Add files via upload
This program will input an undetermined number of student names
and a number of grades for each student. The number of grades is
given by the user. The grades are stored in an array.
Two functions are called for each student.
One function will give the numeric average of their grades.
The other function will give a letter grade to that average.
 Grades are assigned on a 10 point spread.
90-100 A	80-89 B	70-79 C	60-69 D	Below 60 F